Fix circular dependency in Muya
Fix circular dependency between config/index.js and utils/index.js in Muya.
