+++ src/muya/lib/utils/hash.js
@@ -1,0 +1,19 @@
+/**
+ * generate constants map hash, the value is lowercase of the key,
+ * also translate `_` to `-`]
+ */
+export const genUpper2LowerKeyHash = keys => {
+  return keys.reduce((acc, key) => {
+    const value = key.toLowerCase().replace(/_/g, '-')
+    return Object.assign(acc, { [key]: value })
+  }, {})
+}
+
+/**
+ * generate constants map, the value is the key.
+ */
+export const generateKeyHash = keys => {
+  return keys.reduce((acc, key) => {
+    return Object.assign(acc, { [key]: key })
+  }, {})
+}
