+++ src/muya/lib/utils/index.js
@@ -1,8 +1,6 @@
 import runSanitize from './dompurify'
 import { URL_REG, DATA_URL_REG, IMAGE_EXT_REG } from '../config'
-
-const ID_PREFIX = 'ag-'
-let id = 0
+export { getUniqueId, getLongUniqueId } from './random'
 
 const TIMEOUT = 1500
 
@@ -13,10 +11,6 @@ const HTML_TAG_REPLACEMENTS = {
   '"': '&quot;',
   "'": '&#39;'
 }
-
-export const getUniqueId = () => `${ID_PREFIX}${id++}`
-
-export const getLongUniqueId = () => `${getUniqueId()}-${(+new Date()).toString(32)}`
 
 export const isMetaKey = ({ key }) => key === 'Shift' || key === 'Control' || key === 'Alt' || key === 'Meta'
 
@@ -359,26 +353,6 @@ export const wordCount = markdown => {
   all += markdown.length
 
   return { word, paragraph, character, all }
-}
-
-/**
- * [genUpper2LowerKeyHash generate constants map hash, the value is lowercase of the key,
- * also translate `_` to `-`]
- */
-export const genUpper2LowerKeyHash = keys => {
-  return keys.reduce((acc, key) => {
-    const value = key.toLowerCase().replace(/_/g, '-')
-    return Object.assign(acc, { [key]: value })
-  }, {})
-}
-
-/**
- * generate constants map, the value is the key.
- */
-export const generateKeyHash = keys => {
-  return keys.reduce((acc, key) => {
-    return Object.assign(acc, { [key]: key })
-  }, {})
 }
 
 // mixins
